requirements.txt
flask==2.0.2
mysql-connector-python==8.0.28
flask-sqlalchemy==2.5.1
flask-wtf==1.0.0
flask-bcrypt==0.7.1
werkzeug==2.0.0